+++ src/data/products.js
@@ -18,10 +18,62 @@ const siteMyProducts = [
 
 const siteChallengeProducts = [
   {
+    productName: 'Edie homepage',
+    description: (
+      <>
+        devChallenges - Responsive Web 編・課題その7。
+        <br />
+        LP を作成するチャレンジ。
+      </>
+    ),
+    imageUrl: 'img/my-products/devchallenges/responsive-web/edie-homepage.png',
+    url: 'https://h-yoshikawa44.github.io/ch-edie-homepage/',
+    githubUrl: 'https://github.com/h-yoshikawa44/ch-edie-homepage',
+  },
+  {
+    productName: 'Checkout Page',
+    description: (
+      <>
+        devChallenges - Responsive Web 編・課題その6。
+        <br />
+        買い物フォームページを作成するチャレンジ。
+      </>
+    ),
+    imageUrl: 'img/my-products/devchallenges/responsive-web/checkout-page.png',
+    url: 'https://h-yoshikawa44.github.io/ch-checkout-page/',
+    githubUrl: 'https://github.com/h-yoshikawa44/ch-checkout-page',
+  },
+  {
+    productName: 'My Gallery',
+    description: (
+      <>
+        devChallenges - Responsive Web 編・課題その5。
+        <br />
+        ギャラリーページを作成するチャレンジ。
+      </>
+    ),
+    imageUrl: 'img/my-products/devchallenges/responsive-web/my-gallery.png',
+    url: 'https://h-yoshikawa44.github.io/ch-my-gallery/',
+    githubUrl: 'https://github.com/h-yoshikawa44/ch-my-gallery',
+  },
+  {
+    productName: 'Recipe page',
+    description: (
+      <>
+        devChallenges - Responsive Web 編・課題その4。
+        <br />
+        レシピサイトのレシピページを作成するチャレンジ。
+      </>
+    ),
+    imageUrl: 'img/my-products/devchallenges/responsive-web/recipe-page.png',
+    url: 'https://h-yoshikawa44.github.io/ch-recipe-page/',
+    githubUrl: 'https://github.com/h-yoshikawa44/ch-recipe-page',
+  },
+  {
     productName: 'Interior Consultant',
     description: (
       <>
-        devChallenges - Responsive Web 課題その3。
+        devChallenges - Responsive Web 編・課題その3。
         <br />
         インテリアサービスサイトのトップページ + メニューを作成するチャレンジ。
       </>
@@ -35,7 +87,7 @@ const siteChallengeProducts = [
     productName: 'My team page',
     description: (
       <>
-        devChallenges - Responsive Web 課題その2。
+        devChallenges - Responsive Web 編・課題その2。
         <br />
         チームメンバー一覧ページを作成するチャレンジ。
       </>
@@ -48,7 +100,7 @@ const siteChallengeProducts = [
     productName: '404 Not Found',
     description: (
       <>
-        devChallenges - Responsive Web 課題その1。
+        devChallenges - Responsive Web 編・課題その1。
         <br />
         404ページを作成するチャレンジ。
       </>
@@ -93,67 +145,67 @@ const serviceChallengeProducts = [
     productName: 'Random quote generator',
     description: (
       <>
-        devChallenges - Front-end Developer 課題その5。
+        devChallenges - Front-end Developer 編・課題その5。
         <br />
         QuoteGarden API を使用した、名言引用アプリを作るチャレンジ。
       </>
     ),
     imageUrl:
       'img/my-products/devchallenges/front-end/random-quote-generator.png',
-    url: 'https://random-quote-generator-h-yoshikawa44.vercel.app/',
-    githubUrl: 'https://github.com/h-yoshikawa44/random-quote-generator',
+    url: 'https://ch-random-quote-generator-h-yoshikawa44.vercel.app/',
+    githubUrl: 'https://github.com/h-yoshikawa44/ch-random-quote-generator',
   },
   {
     productName: 'Todo app',
     description: (
       <>
-        devChallenges - Front-end Developer 課題その4。
+        devChallenges - Front-end Developer 編・課題その4。
         <br />
         LocalStorage を使用したシンプルな Todo アプリを作るチャレンジ。
       </>
     ),
     imageUrl: 'img/my-products/devchallenges/front-end/todo-app.png',
-    url: 'https://todo-app-h-yoshikawa44.vercel.app',
-    githubUrl: 'https://github.com/h-yoshikawa44/todo-app',
+    url: 'https://ch-todo-app-h-yoshikawa44.vercel.app',
+    githubUrl: 'https://github.com/h-yoshikawa44/ch-todo-app',
   },
   {
     productName: 'Windbnb',
     description: (
       <>
-        devChallenges - Front-end Developer 課題その3。
+        devChallenges - Front-end Developer 編・課題その3。
         <br />
         簡易的な検索機能を持つ、ミニマムな Airbnb クローンを作るチャレンジ。
       </>
     ),
     imageUrl: 'img/my-products/devchallenges/front-end/windbnb.png',
-    url: 'https://windbnb-h-yoshikawa44.vercel.app',
-    githubUrl: 'https://github.com/h-yoshikawa44/windbnb',
+    url: 'https://ch-windbnb-h-yoshikawa44.vercel.app',
+    githubUrl: 'https://github.com/h-yoshikawa44/ch-windbnb',
   },
   {
     productName: 'Input component',
     description: (
       <>
-        devChallenges - Front-end Developer 課題その2。
+        devChallenges - Front-end Developer 編・課題その2。
         <br />
         汎用的なボタンコンポーネントと、そのカタログページを作るチャレンジ。
       </>
     ),
     imageUrl: 'img/my-products/devchallenges/front-end/input-component.png',
-    url: 'https://input-component-h-yoshikawa44.vercel.app',
-    githubUrl: 'https://github.com/h-yoshikawa44/input-component',
+    url: 'https://ch-input-component-h-yoshikawa44.vercel.app',
+    githubUrl: 'https://github.com/h-yoshikawa44/ch-input-component',
   },
   {
     productName: 'Button component',
     description: (
       <>
-        devChallenges - Front-end Developer 課題その1。
+        devChallenges - Front-end Developer 編・課題その1。
         <br />
         汎用的なボタンコンポーネントと、そのカタログページを作るチャレンジ。
       </>
     ),
     imageUrl: 'img/my-products/devchallenges/front-end/button-component.png',
-    url: 'https://button-component-h-yoshikawa44.vercel.app',
-    githubUrl: 'https://github.com/h-yoshikawa44/button-component',
+    url: 'https://ch-button-component-h-yoshikawa44.vercel.app',
+    githubUrl: 'https://github.com/h-yoshikawa44/ch-button-component',
   },
 ];