+++ src/pages/index.js
@@ -94,6 +94,9 @@ function Home() {
                       <br />
                       最近、メインで使用しているものは TypeScript × React、時に
                       Next.js です。
+                      <br />
+                      また、基礎の地盤固めとして HTML・CSS・JavaScript
+                      の再勉強をしていたりもします。
                     </p>
                   </div>
                   <div className="card__footer">