+++ CHANGELOG.md
@@ -1,9 +1,31 @@
 # CHANGELOG
 h-yoshikawa44's Portfolio の大まかな機能変化記録
 
+## 2021-11-25
+
+### Changed
+- MyProductsのカードの影と説明文部分の高さ調整(#27)
+
+### Updated
+- ポートフォリオ更新 - 2021/11 後期(#27)
+  - MyProducts
+    - Recipe page（devChallenges - Responsive Web 編）
+    - My Gallery（devChallenges - Responsive Web 編）
+    - Checkout Page（devChallenges - Responsive Web 編）
+    - Edie homepage（devChallenges - Responsive Web 編）
+- Node.js：14.17.3 → 16.13.0
+
+### Other
+- ESLint・Prettier の Pre Commit 設定(#27)
+
 ## 2021-11-01
 ### Updated
 - ポートフォリオ更新 - 2021/11(#25)
+  - MyProducts
+    - Change Of Pace（※ブログ自体は以前よりあったが、このページには載せてなかった）
+    - 404 Not Found（devChallenges - Responsive Web 編）
+    - My team page（devChallenges - Responsive Web 編）
+    - Interior Consultant（devChallenges - Responsive Web 編）
 - Node.js：14.16.0 → 14.17.3
 
 ## 2021-09-28
@@ -13,6 +35,8 @@ h-yoshikawa44's Portfolio の大まかな機能変化記録
 ## 2021-08-18
 ### Updated
 - ポートフォリオ更新 - 2021/08 中間(#21)
+  - アカウントリンクに Wantedly を追加
+  - web1week 作品の技術構成を更新（TypeScript 化したので）
 
 ## 2021-08-01
 ### Changed
@@ -26,6 +50,8 @@ h-yoshikawa44's Portfolio の大まかな機能変化記録
 
 ### Updated
 - ポートフォリオ更新 - 2021/08(#17)
+  - MyProducts：
+    - Random quote generator（devChallenges - Front-end Developer 編）
 
 ## 2021-07-04
 ### TIL Docs・Removed
@@ -34,6 +60,11 @@ h-yoshikawa44's Portfolio の大まかな機能変化記録
 ## 2021-06-30
 ### Updated
 - ポートフォリオ更新 - 2021/07(#13)
+  - MyProducts：
+    - Button component（devChallenges - Front-end Developer 編）
+    - Input component（devChallenges - Front-end Developer 編）
+    - Windbnb（devChallenges - Front-end Developer 編）
+    - TodoApp（devChallenges - Front-end Developer 編）
 
 ## 2021-05-14
 ### Updated